Move WriteOperation to core-load and remove duplicates from modern connectors.
