+++ airbyte-cdk/bulk/core/load/src/test/kotlin/io/airbyte/cdk/load/write/WriteOperationTest.kt
@@ -2,7 +2,7 @@
  * Copyright (c) 2026 Airbyte, Inc., all rights reserved.
  */
 
-package io.airbyte.integrations.destination.snowflake.cdk
+package io.airbyte.cdk.load.write
 
 import io.airbyte.cdk.load.dataflow.DestinationLifecycle
 import io.mockk.every
@@ -10,12 +10,12 @@ import io.mockk.mockk
 import io.mockk.verify
 import org.junit.jupiter.api.Test
 
-internal class WriteOperationV2Test {
+class WriteOperationTest {
 
     @Test
     fun testWriteOperation() {
         val destinationLifecycle = mockk<DestinationLifecycle> { every { run() } returns Unit }
-        val writeOperation = WriteOperationV2(destinationLifecycle)
+        val writeOperation = WriteOperation(destinationLifecycle)
         writeOperation.execute()
         verify(exactly = 1) { destinationLifecycle.run() }
     }
