--- airbyte-integrations/connectors/destination-gcs-data-lake/src/main/kotlin/io/airbyte/integrations/destination/gcs_data_lake/cdk/WriteOperationV2.kt
@@ -1,27 +1,0 @@
-/*
- * Copyright (c) 2026 Airbyte, Inc., all rights reserved.
- */
-
-package io.airbyte.integrations.destination.gcs_data_lake.cdk
-
-import io.airbyte.cdk.Operation
-import io.airbyte.cdk.load.dataflow.DestinationLifecycle
-import io.github.oshai.kotlinlogging.KotlinLogging
-import io.micronaut.context.annotation.Primary
-import io.micronaut.context.annotation.Requires
-import jakarta.inject.Singleton
-
-@Primary
-@Singleton
-@Requires(property = Operation.PROPERTY, value = "write")
-class WriteOperationV2(
-    private val d: DestinationLifecycle,
-) : Operation {
-    private val log = KotlinLogging.logger {}
-
-    override fun execute() {
-        log.info { "Running new pipe..." }
-        d.run()
-        log.info { "New pipe complete :tada:" }
-    }
-}
